+++ rotkehlchen/chain/ethereum/structures.py
@@ -72,10 +72,14 @@ class AaveEvent:
     def __eq__(self, other: Any) -> bool:
         return hash(self) == hash(other)
 
+    def __str__(self) -> str:
+        """Used in DefiEvent processing during accounting"""
+        return f'Aave {self.event_type} event'
+
 
 @dataclasses.dataclass(init=True, repr=True, eq=False, order=False, unsafe_hash=False, frozen=True)
 class AaveSimpleEvent(AaveEvent):
-    """A simple event of the Aave protocol. Deposit or withdrawal"""
+    """A simple event of the Aave protocol. Deposit, withdrawal or interest"""
     asset: Asset
     value: Balance
 
@@ -137,7 +141,7 @@ class AaveRepayEvent(AaveSimpleEvent):
 
 @dataclasses.dataclass(init=True, repr=True, eq=False, order=False, unsafe_hash=False, frozen=True)
 class AaveLiquidationEvent(AaveEvent):
-    """A simple event of the Aave protocol. Deposit or withdrawal"""
+    """An aave liquidation event. You gain the principal and lose the collateral."""
     collateral_asset: Asset
     collateral_balance: Balance
     principal_asset: Asset
@@ -310,6 +314,10 @@ class YearnVaultEvent:
             'tx_hash': self.tx_hash,
             'log_index': self.log_index,
         }
+
+    def __str__(self) -> str:
+        """Used in DefiEvent processing during accounting"""
+        return f'Yearn vault {self.event_type}'
 
 
 class YearnVault(NamedTuple):